+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [Tool] Update `SwiftFormat` Swift Package from v0.56.2 to v0.56.4
 - Optional accessibility identifiers for items of radio picker and checkbox picker components
+- Switch component tokens (tokens library v1.1.0) (Orange-OpenSource/ouds-ios#664)
 - Checkbox component tokens (tokens library v1.1.0) (Orange-OpenSource/ouds-ios#664)
 - Control item component tokens (tokens library v1.1.0) (Orange-OpenSource/ouds-ios#664)
 - Link component tokens (tokens library v1.1.0) (Orange-OpenSource/ouds-ios#664)